Basic Tetromino movement + edge detection

+++ src/game/grid.rs
@@ -25,8 +25,8 @@ impl Grid {
         for y in 0..4 {
             for x in 0..4 {
                 if let Some(ref block) = tetromino.blocks[y][x] {
-                    let x = (tetromino.x + x as f64) as usize;
-                    let y = (tetromino.y + y as f64) as usize;
+                    let x = (tetromino.x + x as i32) as usize;
+                    let y = (tetromino.y + y as i32) as usize;
 
                     if x >= 10 || y >= 22 {
                         continue;
